Remove unused @MAN2TCLFLAGS@ Move <limits.h> include from tclInt.h to tclWinPort.h, and eliminate unneeded <stdlib.h>, <stdio.h> and <string.h>, which are already in tclInt.h Move "tclInt.h" from regcustom.h up to regex.h. tclAlloc.c: Unneeded <stdio.h> include tclExecute.c: Fix gcc warning: comparison between signed and unsigned

+++ generic/tclExecute.c
@@ -5081,7 +5081,7 @@ TclExecuteByteCode(
 		     * Quickly force large right shifts to 0 or -1.
 		     */
 
-		    if (l2 >= CHAR_BIT*sizeof(long)) {
+		    if (l2 >= (long)(CHAR_BIT*sizeof(long))) {
 			/*
 			 * We assume that INT_MAX is much larger than the
 			 * number of bits in a long. This is a pretty safe
